Improve typing (#516)

+++ tests/test_logo_client.py
@@ -21,42 +21,44 @@ class TestLogoClient(unittest.TestCase):
     process = None
 
     @classmethod
-    def setUpClass(cls):
+    def setUpClass(cls) -> None:
         cls.process = Process(target=mainloop)
         cls.process.start()
         time.sleep(2)  # wait for server to start
 
     @classmethod
-    def tearDownClass(cls):
+    def tearDownClass(cls) -> None:
+        if cls.process is None:
+            return
         cls.process.terminate()
         cls.process.join(1)
         if cls.process.is_alive():
             cls.process.kill()
 
-    def setUp(self):
+    def setUp(self) -> None:
         self.client = snap7.logo.Logo()
         self.client.connect(ip, rack, slot, tcpport)
 
-    def tearDown(self):
+    def tearDown(self) -> None:
         self.client.disconnect()
         self.client.destroy()
 
-    def test_read(self):
+    def test_read(self) -> None:
         vm_address = "V40"
         value = 50
         self.client.write(vm_address, value)
         result = self.client.read(vm_address)
         self.assertEqual(value, result)
 
-    def test_write(self):
+    def test_write(self) -> None:
         vm_address = "V20"
         value = 8
         self.client.write(vm_address, value)
 
-    def test_get_connected(self):
+    def test_get_connected(self) -> None:
         self.client.get_connected()
 
-    def test_set_param(self):
+    def test_set_param(self) -> None:
         values = (
             (snap7.types.PingTimeout, 800),
             (snap7.types.SendTimeout, 15),
@@ -71,7 +73,7 @@ class TestLogoClient(unittest.TestCase):
 
         self.assertRaises(Exception, self.client.set_param, snap7.types.RemotePort, 1)
 
-    def test_get_param(self):
+    def test_get_param(self) -> None:
         expected = (
             (snap7.types.RemotePort, tcpport),
             (snap7.types.PingTimeout, 750),
@@ -106,10 +108,10 @@ class TestClientBeforeConnect(unittest.TestCase):
     Test suite of items that should run without an open connection.
     """
 
-    def setUp(self):
+    def setUp(self) -> None:
         self.client = snap7.client.Client()
 
-    def test_set_param(self):
+    def test_set_param(self) -> None:
         values = (
             (snap7.types.RemotePort, 1102),
             (snap7.types.PingTimeout, 800),
